Improve run_action module
Now inline() automatically converts known variables (such as property
maps) and defaults to boost::python objects, instead of scxx objects.

The code now is only generated for the current filtered, reversed and/or
directed status of the graph, reducing compile time and binary size.

Edge modification (remove and add) is now protected by a GraphWrap
class, which takes care of the edge index housekeeping, which does not
need to be done by hand anymore.

The function is also no longer bound to one graph, and can take an
arbitrary number of variables of known or unknown types, including
graphs.