Fix spurious parallel edge creation in random_rewire()
In some circumstances, the test for parallel edges would fail to detect
one, if it involved another "new" edge and one of the current ones being
rewired.

This commit also removes the requirement that edge indexes be continuous
in the range [0, num_edges(g)), which is not in general the case.