Speed up motifs() and motif_significance()
In both functions we can save some time by hashing the subgraphs found,
according to their degree distribution signature (instead of number of
edges). This avoids a large number of useless exactness or isomorphism
comparisons.

This also removes the "seed" parameter, which is redundant to
the numpy.random.seed() function.