Refactor random_rewire()
This extensively modifies the random_rewire() algorithm, so that only
either the source or the edge of each edge is rewired (not both, as
previously), and no parallel edges are created during the algorithm (if
desired).

The new version is much faster, and never gets stuck. However, more than
one run may be necessary in order to obtain a uniform shuffling.