Improve price_network() seeding semantics
Now, if 'm' is larger than the seed graph, it is set to the size of the
seed graph, and is increased progressively until it matches the desired
value. This removes the need to start with a random seed graph of the
appropriate size.

This also includes modifications which make the code more robust
against parameter choices which don't make much sense, e.g. lead to
negative probabilites.